Make everything in a single example
